Update bazel file monitor script

+++ examples/goldilocks_models/cluster_scripts/monitor_bazel_build_file_age.py
@@ -14,9 +14,11 @@ def file_age(filepath):
 file_path = '~/bazel_is_building'
 
 while True:
-  seconds = file_age(file_path)
-  if seconds / 3600 > 2:
-    os.remove(file_path)
+  if os.path.exists(file_path):
+    seconds = file_age(file_path)
+    if seconds / 3600 > 2:
+      print("file is too old, deleting")
+      os.remove(file_path)
 
   time.sleep(60)
